finding-different-tools-sever/requirements.txt
Flask
ultralytics
Pillow
numpy
opencv-python-headless
scikit-image
imutils
flask-cors